fix/initapp_in_main.ts
